let the system style choose which color to substitute «^7» with
^7 (white) is the default foreground text color in Unvanquished where
black is the dominant background color, which doesn't render well with
user-configurable background colors (as in system styles). therefore,
ignore «^7» sequences in server and player names to let the OS choose
the optimal foreground font and eradicate readability problems with
white-on-white (white-on-light).
also, a few minor optimizations and style improvements.